pyproject.toml
[project]
name = "atmio"
version = "0.1.0"
description = ""
authors = [
    {name = "jakubjagielski8@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0.1,<7.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
